Rename lib to dist
