+++ packages/react-cosmos-redux-proxy/package.json
@@ -5,7 +5,8 @@
   "repository": "https://github.com/react-cosmos/react-cosmos/tree/master/packages/react-cosmos-redux-proxy",
   "license": "MIT",
   "files": [
-    "lib"
+    "dist",
+    "index.js"
   ],
   "main": "index.js",
   "dependencies": {
