+++ packages/react-cosmos-scripts/bin/upgrade-fixtures.js
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
 
-const upgradeFixtures = require('../lib/upgrade-fixtures').default;
+const upgradeFixtures = require('../dist/upgrade-fixtures').default;
 
 upgradeFixtures();
