+++ packages/react-cosmos-shared/package.json
@@ -5,7 +5,10 @@
   "repository": "https://github.com/react-cosmos/react-cosmos/tree/master/packages/react-cosmos-shared",
   "license": "MIT",
   "files": [
-    "lib"
+    "dist",
+    "index.js",
+    "react.js",
+    "server.js"
   ],
   "main": "index.js",
   "dependencies": {
