+++ packages/react-cosmos/bin/cosmos-export.js
@@ -4,6 +4,6 @@
 process.env.BABEL_ENV = process.env.BABEL_ENV || 'production';
 process.env.NODE_ENV = process.env.NODE_ENV || 'production';
 
-const startExport = require('../lib/server/export').default;
+const startExport = require('../dist/server/export').default;
 
 startExport();
