+++ packages/react-cosmos/bin/cosmos.js
@@ -4,6 +4,6 @@
 process.env.BABEL_ENV = process.env.BABEL_ENV || 'development';
 process.env.NODE_ENV = process.env.NODE_ENV || 'development';
 
-const startServer = require('../lib/server/server').default;
+const startServer = require('../dist/server/server').default;
 
 startServer();
