+++ packages/react-cosmos/bin/print-fixture-files.js
@@ -1,9 +1,9 @@
 #!/usr/bin/env node
 
 // Do this as the first thing so that any code reading it knows the right env.
-process.env.BABEL_ENV = 'development';
-process.env.NODE_ENV = 'development';
+process.env.BABEL_ENV = process.env.BABEL_ENV || 'development';
+process.env.NODE_ENV = process.env.NODE_ENV || 'development';
 
-const printFixtureFiles = require('../lib/server/print-fixture-files').default;
+const printFixtureFiles = require('../dist/server/print-fixture-files').default;
 
 printFixtureFiles();
